Write book cover output to temporary files converted afterwards

+++ src/main.rs
@@ -4,6 +4,7 @@ use std::io::Write;
 
 use std::env;
 use sha256::digest_file;
+use std::path::Path;
 
 fn main() {
     let mut extract_cover = true;
@@ -35,16 +36,20 @@ fn main() {
         let title = doc.mdata("title").unwrap();
 
         // Cover
-        let mut cover_path = main_path.clone() + &digest_file(&epub_file).unwrap().to_string();
-        if extract_cover == true {
-            let cover_data = doc.get_cover();
-            if cover_data.is_ok() {
-                let f = fs::File::create(cover_path.clone());
-                let mut f = f.unwrap();
-                f.write_all(&cover_data.unwrap()).unwrap();
-            }
-            else {
-                cover_path = "".to_string();
+        let file_digest = digest_file(&epub_file).unwrap().to_string();
+        let mut cover_path = main_path.clone() + &file_digest + ".t";
+        let cover_path_converted = main_path.clone() + &file_digest;
+        if !Path::new(&cover_path_converted).exists() {
+            if extract_cover == true {
+                let cover_data = doc.get_cover();
+                if cover_data.is_ok() {
+                    let f = fs::File::create(cover_path.clone());
+                    let mut f = f.unwrap();
+                    f.write_all(&cover_data.unwrap()).unwrap();
+                }
+                else {
+                    cover_path = "".to_string();
+                }
             }
         }
 
@@ -76,6 +81,5 @@ fn main() {
         }
         main_string.push_str(&new_json);
     }
-    main_string.push_str("]}");
     print!("{}", main_string);
 }
